# Enable `RestoreUseStaticGraphEvaluation` to speed up `Restore` stage