--- a/tracer/src/Datadog.InstrumentedAssemblyGenerator/Datadog.InstrumentedAssemblyGenerator.csproj
+++ b/tracer/src/Datadog.InstrumentedAssemblyGenerator/Datadog.InstrumentedAssemblyGenerator.csproj
@@ -6,7 +6,6 @@
     <Copyright>Copyright 2022 Datadog, Inc.</Copyright>
     
     <LangVersion>latest</LangVersion>
-    <TargetFramework>netstandard2.0</TargetFramework>
     <RunAnalyzersDuringBuild>False</RunAnalyzersDuringBuild>
     <AnalysisLevel>none</AnalysisLevel>
     <RunAnalyzersDuringLiveAnalysis>False</RunAnalyzersDuringLiveAnalysis>