--- a/tracer/src/Datadog.Trace.Coverage.collector/Datadog.Trace.Coverage.collector.csproj
+++ b/tracer/src/Datadog.Trace.Coverage.collector/Datadog.Trace.Coverage.collector.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
         <ImplicitUsings>false</ImplicitUsings>
         <Nullable>enable</Nullable>
         <SuppressTfmSupportBuildWarnings>true</SuppressTfmSupportBuildWarnings>