## Summary of changes

This enables the usage of `RestoreUseStaticGraphEvaluation` during the NuGet `Restore` stage to speed up the build (slightly).

Doc: https://learn.microsoft.com/en-us/nuget/reference/msbuild-targets#restoring-with-msbuild-static-graph-evaluation

The restore usually takes me ~45 seconds to ~60 seconds per build, this change brings that down to ~10 seconds to ~25 seconds.

## Reason for change

Just trying various things to speed up the build process.

## Implementation details

Add `RestoreUseStaticGraphEvaluation` as a property to the `dotnet` call.

No longer use `NuGetTasks` for Windows restoring - unsure if that is important or not?

Had to remove specific `<TargetFramework>netstandard2.0</TargetFramework>` as it was causing the following build error:

```
C:\Program Files\dotnet\sdk\8.0.103\NuGet.RestoreEx.targets(19,5): error : Invalid restore input. Duplicate frameworks found: 'netstandard2.0, netstandard2.0, netstandard2.0, netstandard2.0'. Input files: \Datadog.Trace.Coverage.collector\Datadog.Trace.Coverage.collector.csproj.
```

Had to ignore some `null` issues with `Coverage.Collector`

I also noticed this message that pops up with the `Restore` now and after looking into what it means I don't really know what to do with it, but I can't tell if it is an issue:

```
The solution contains '22' project(s) 'Solution Items,src,test,test-applications,regression,dependency-libs,integrations,aspnet,dependency-libs,build,artifacts,benchmarks,instrumentation,crank,shared,shared,security,aspnet,azure-functions,Samples.Shared,debugger,dependency-libs' that are not known to MSBuild. Ensure that all projects are known to be MSBuild before running restore on the solution.
```

I've also noticed that there is a `FastDevLoop` option that was added to our Nuke build process that completely bypasses the Restore stage, as an alternative we could just do use I suppose and close this PR.

## Test coverage

- If it works on the CI is the test.

## Other details
<!-- Fixes #{issue} -->

<!--  ⚠️ Note: where possible, please obtain 2 approvals prior to merging. Unless CODEOWNERS specifies otherwise, for external teams it is typically best to have one review from a team member, and one review from apm-dotnet. Trivial changes do not require 2 reviews. -->