@bouwkast 2024-04-09
> I have a strong suspicion this isn't going to work in CI, and has conflicts with the v3 branch, but 🤞 😅

Would it be best to just close this then?