@bouwkast 2024-04-09
> > Would it be best to just close this then?
> 
> Given that it looks like CI is unhappy, unfortunately, I suspect so. I'm definitely going to try to work this in as I try to refactor the build in general though 🙂

Well it worked on my machine 😆 


Closing this PR as it seems that this would require a decent bit of work to get working on CI and isn't a main focus at the moment.